refactor(eve): separate tool facade from implementation
