+++ packages/eve/src/tools/bash.ts
@@ -1,0 +1,43 @@
+import { z } from "#compiled/zod/index.js";
+
+import { executeBashOnSandbox, type BashInput } from "#execution/sandbox/bash-tool.js";
+import { defineTool, type ToolDefinition } from "#public/definitions/tool.js";
+
+/**
+ * Input schema for the provided `bash` tool.
+ */
+export const BASH_INPUT_SCHEMA = z.strictObject({
+  command: z.string().describe("The shell command to execute."),
+});
+
+/**
+ * Output schema for the provided `bash` tool.
+ */
+export const BASH_OUTPUT_SCHEMA = z.strictObject({
+  exitCode: z.number(),
+  stderr: z.string(),
+  stdout: z.string(),
+  truncated: z.boolean(),
+});
+
+export type BashToolInput = z.infer<typeof BASH_INPUT_SCHEMA>;
+export type BashToolOutput = z.infer<typeof BASH_OUTPUT_SCHEMA>;
+
+/**
+ * Framework-owned executors stay statically imported so hosted server bundles
+ * can trace and rewrite them into deployable output chunks.
+ *
+ * These modules are only used by the Nitro-hosted runtime path. Their deeper
+ * sandbox dependencies remain lazily loaded inside the execution layer, so the
+ * top-level import here does not force those backends to initialize eagerly.
+ */
+export const bash: ToolDefinition<BashToolInput, BashToolOutput> = defineTool({
+  description: "Execute a shell command in the shared workspace environment.",
+  async execute(input, ctx) {
+    return await executeBashOnSandbox(await ctx.getSandbox(), input as BashInput);
+  },
+  inputSchema: BASH_INPUT_SCHEMA,
+  outputSchema: BASH_OUTPUT_SCHEMA,
+});
+
+export default bash;
